Move CLCalculator tests and apply black
